Rename RedisConfig and fix namespace
